#11 add SharedService Class
